Mott's Apple Juice aight thanks

+++ Askew/core/src/askew/entity/sloth/SlothModel.java
@@ -122,6 +122,8 @@ public class SlothModel extends ComplexObstacle  {
     private boolean setLastGrabX;
     private float lastGrabX;
     private transient boolean dismembered;
+    private boolean didOneArmCheck;
+    private boolean waitingForSafeRelease;
     private transient boolean tutorial = false;
 
     /**
@@ -158,7 +160,7 @@ public class SlothModel extends ComplexObstacle  {
 
     private static final float HAND_YOFFSET    = 0;
 
-    private static final float BODY_HEIGHT = 1.4f;
+    private static final float BODY_HEIGHT = 1.8f;
     private static final float BODY_WIDTH = 1.8f * (489f / 835f);
 
     private static final float ARM_WIDTH = 1.75f;
@@ -646,6 +648,7 @@ public class SlothModel extends ComplexObstacle  {
     }
 
     public void setLeftGrab(boolean leftGrab) {
+        if (controlMode == CONTROLS_ONE_ARM && !didOneArmCheck) return;
         if (movementMode == GRAB_ORIGINAL)
             this.leftGrab = leftGrab;
         else if (movementMode == GRAB_REVERSE)
@@ -659,22 +662,28 @@ public class SlothModel extends ComplexObstacle  {
             }
             this.leftGrab = leftGrabbing;
         }
+        didOneArmCheck = false;
     }
 
     public void setRightGrab(boolean rightGrab) {
-        if (movementMode == GRAB_ORIGINAL)
-            this.rightGrab = rightGrab;
-        else if (movementMode == GRAB_REVERSE)
-            this.rightGrab = !rightGrab;
-        else if (movementMode == GRAB_TOGGLE && rightGrab) {
-            if (!rightGrabbing) {
-                rightGrabbing = true;
+        if (controlMode != CONTROLS_ONE_ARM || didOneArmCheck) {
+            if (movementMode == GRAB_ORIGINAL)
+                this.rightGrab = rightGrab;
+            else if (movementMode == GRAB_REVERSE)
+                this.rightGrab = !rightGrab;
+            else if (movementMode == GRAB_TOGGLE && rightGrab) {
+                if (!rightGrabbing) {
+                    rightGrabbing = true;
+                }
+                else {
+                    rightGrabbing = false;
+                }
+                this.rightGrab = rightGrabbing;
             }
-            else {
-                rightGrabbing = false;
-            }
-            this.rightGrab = rightGrabbing;
+        } else {
+            setOneGrab(rightGrab);
         }
+        didOneArmCheck = false;
     }
 
     public void setLeftStickPressed(boolean leftStickPressed) {
@@ -698,6 +707,7 @@ public class SlothModel extends ComplexObstacle  {
     }
 
     public void grab(World world, Body target, boolean leftHand) {
+        if (didSafeGrab) return;
         Joint grabJoint;
         Joint otherGrabJoint;
         RevoluteJointDef grabJointDef;
@@ -753,6 +763,7 @@ public class SlothModel extends ComplexObstacle  {
     }
 
     public void releaseLeft(World world) {
+        if (didSafeGrab) return;
         if (leftGrabJoint != null) {
             if (movementMode != GRAB_TOGGLE || !leftGrabbing) world.destroyJoint(leftGrabJoint);
             leftCanGrabOrIsGrabbing = false;
@@ -761,6 +772,7 @@ public class SlothModel extends ComplexObstacle  {
     }
 
     public void releaseRight(World world) {
+        if (didSafeGrab) return;
         if (rightGrabJoint != null) {
             if (movementMode != GRAB_TOGGLE || !rightGrabbing) world.destroyJoint(rightGrabJoint);
             leftCanGrabOrIsGrabbing = true;
@@ -937,13 +949,16 @@ public class SlothModel extends ComplexObstacle  {
      * @param b
      */
     public void setOneGrab(boolean b) {
-        grabbedEntity = false;
-        if (leftCanGrabOrIsGrabbing) {
-            rightGrab = false;
-            setLeftGrab(b);
-        } else {
-            leftGrab = false;
-            setRightGrab(b);
+        if (!didSafeGrab) {
+            didOneArmCheck = true;
+            grabbedEntity = false;
+            if (leftCanGrabOrIsGrabbing) {
+                rightGrab = false;
+                setLeftGrab(b);
+            } else {
+                leftGrab = false;
+                setRightGrab(b);
+            }
         }
     }
 
@@ -957,21 +972,26 @@ public class SlothModel extends ComplexObstacle  {
      */
     public void setSafeGrab(boolean leftButtonPressed, Body leftCollisionBody, Body rightCollisionBody, World world) {
         grabbedEntity = false;
+        didSafeGrab = false;
+        if (waitingForSafeRelease && leftButtonPressed) {
+            return;
+        }
+        waitingForSafeRelease = false;
         if (leftButtonPressed) {
-            if (!didSafeGrab) {
-                if (isActualLeftGrab()) {
-                    if (rightCollisionBody != null) {
-                        releaseLeft(world);
-                        grab(world, rightCollisionBody, false);
+            if (isActualLeftGrab()) {
+                if (rightCollisionBody != null) {
+                    releaseLeft(world);
+                    grab(world, rightCollisionBody, false);
+                    didSafeGrab = true;
+                    waitingForSafeRelease = true;
+                }
+            } else {
+                if (isActualRightGrab()) {
+                    if (leftCollisionBody != null) {
+                        releaseRight(world);
+                        grab(world, leftCollisionBody, true);
                         didSafeGrab = true;
-                    }
-                } else {
-                    if (isActualRightGrab()) {
-                        if (leftCollisionBody != null) {
-                            releaseRight(world);
-                            grab(world, leftCollisionBody, true);
-                            didSafeGrab = true;
-                        }
+                        waitingForSafeRelease = true;
                     }
                 }
             }
@@ -990,6 +1010,8 @@ public class SlothModel extends ComplexObstacle  {
             world.destroyJoint(jointB);
             jointA = null;
             jointB = null;
+            bodies.get(0).setFixedRotation(false);
+            bodies.get(0).getBody().applyAngularImpulse(0.5f,true);
             for (Obstacle b : bodies) {
                 b.getFilterData().categoryBits = 0;
                 b.getBody().applyForceToCenter((float)Math.random()*110 - 55,(float)Math.random()*110 - 55,true);
